requirements.txt
Pillow
screeninfo==0.6.7
